Tab and HelmTab use hooks, Removing unnecessary render in HelmColumns.ts
