+++ frontend/src/types/index.ts
@@ -1,4 +1,4 @@
 export type {MenuItem} from './menuItem.ts'
 export type {HelmRelease, HelmReleaseList, Status, HelmReleaseHistory, HelmReleaseHistoryList, ReleaseNameRollbackBody} from './helmTypes.ts'
-export type {HelmDataSourceItem, HelmColumnType} from './helmTypes.ts'
+export type {ResourceDataSourceItem, ResourceColumnType, HelmDataSourceItem, HelmColumnType} from './tabTypes.ts'
 export type {HelmModalProps} from './modalProps.ts'
